♻️ Separate concerns
Now that we know that the conversion is functional, we can separate the actions so that every class has one purpose.
I introduced the builder classes which are made to be called to create the instances of users and profiles. In a next step, i could try to prevent the initialize of these last instances outside of the builders.
I also separate the json importer and the csv generator for better reading.
Clause guards have been added to protect the future database from invalid entries.